use my own buf thing

+++ src/dns.rs
@@ -29,7 +29,6 @@ pub enum QuestionType {
     ALL         = 0xff
 }
 
-
 impl QuestionType {
     fn new(i: u16) -> Option<QuestionType> {
         use self::QuestionType::*;
@@ -55,7 +54,7 @@ pub enum QuestionClass {
 
 pub const LABEL_MAX_LENGTH: usize = 63;
 pub const NAMES_MAX_LENGTH: usize = 255;
-pub const UDP_MAX_LENGTH: u16  = 512;
+pub const UDP_MAX_LENGTH: usize  = 512;
 
 impl QuestionClass {
     fn new(i: u16) -> Option<QuestionClass> {
